chore(web): rename SiNavbarButton to SiBarButton
Since SiNavbarButton is used in the ChangeSetTabPanel, it is now in the
status bar too. This prompted an uneventful name change.
