+++ app/web/src/organisms/SiThemeSwitcher.vue
@@ -1,6 +1,6 @@
 <template>
   <!-- FIXME(nick,theo): dropdown-classes needs to be removed in favor of the dropdown knowing whether or not it is offscreen. -->
-  <SiNavbarButton tooltip-text="Change theme" dropdown-classes="-right-12">
+  <SiBarButton tooltip-text="Change theme" dropdown-classes="-right-12">
     <div v-if="lightmode">
       <MoonIcon class="w-6" />
     </div>
@@ -28,7 +28,7 @@
         Dark theme
       </SiDropdownItem>
     </template>
-  </SiNavbarButton>
+  </SiBarButton>
 </template>
 
 <script setup lang="ts">
@@ -36,7 +36,7 @@ import { ThemeService } from "@/service/theme";
 import SiDropdownItem from "@/atoms/SiDropdownItem.vue";
 import MoonIcon from "@/atoms/CustomIcons/MoonIcon.vue";
 import SunIcon from "@/atoms/CustomIcons/SunIcon.vue";
-import SiNavbarButton from "@/molecules/SiNavbarButton.vue";
+import SiBarButton from "@/molecules/SiBarButton.vue";
 import { Theme } from "@/observable/theme";
 import { computed } from "vue";
 import { refFrom } from "vuse-rx/src";
